Created Endpoint Search cards LessThanEqual Limit

+++ mscartoes/src/main/java/com/claudiocarige/mscartoes/infra/repository/CardPersonalRepository.java
@@ -3,8 +3,9 @@ package com.claudiocarige.mscartoes.infra.repository;
 import com.claudiocarige.mscartoes.domain.CardPersonal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CardPersonalRepository extends JpaRepository<CardPersonal, Long> {
-    List<CardPersonal> findByIncomeLessThanEqual(Long income);
+    List<CardPersonal> findByIncomeLessThanEqual(BigDecimal income);
 }
